+++ .github/actions/download-pr-data-artifact/action.yaml
@@ -29,7 +29,7 @@ runs:
           });
           let fs = require('fs');
           fs.writeFileSync(`${process.env.GITHUB_WORKSPACE}/pr_number.zip`, Buffer.from(download.data));
-    - name: 'Unzip artifact'
+    - name: "Unzip artifact"
       shell: bash
       run: unzip pr_number.zip
     - name: Set PR number
