+++ .github/workflows/functional-tests-approval.yaml
@@ -1,4 +1,4 @@
-name: 'Approve Functional Tests'
+name: "Approve Functional Tests"
 on:
   pull_request:
     branches:
@@ -7,11 +7,12 @@ on:
       - release/*
 jobs:
   approve-functional-tests-run:
-    name: 'Approve Functional Tests'
+    name: "Approve Functional Tests"
     runs-on: ubuntu-latest
     environment: functional-tests
     steps:
-    - name: Use custom actions
-      uses: actions/checkout@v4
-    - name: Save PR number
-      uses: ./.github/actions/save-pr-as-artifact
+      - name: Checkout Radius repository
+        uses: actions/checkout@v4
+
+      - name: Save PR number
+        uses: ./.github/actions/save-pr-as-artifact
