+++ test/functional-portable/datastoresrp/noncloud/resources/testdata/datastoresrp-resources-redis-default-recipe.bicep
@@ -1,28 +1,21 @@
 import radius as radius
 
-param scope string = resourceGroup().id
-
 param registry string
 
 param version string
 
 resource env 'Applications.Core/environments@2023-10-01-preview' = {
-  name: 'dsrp-resources-env-recipe-env'
+  name: 'dsrp-resources-env-default-recipe-env'
   location: 'global'
   properties: {
     compute: {
       kind: 'kubernetes'
       resourceId: 'self'
-      namespace: 'dsrp-resources-env-recipe-env'
-    }
-    providers: {
-      azure: {
-        scope: scope
-      }
+      namespace: 'dsrp-resources-env-default-recipe-env'
     }
     recipes: {
       'Applications.Datastores/redisCaches': {
-        rediscache: {
+        default: {
           templateKind: 'bicep'
           templatePath: '${registry}/test/testrecipes/test-bicep-recipes/redis-recipe-value-backed:${version}'
         }
@@ -32,27 +25,24 @@ resource env 'Applications.Core/environments@2023-10-01-preview' = {
 }
 
 resource app 'Applications.Core/applications@2023-10-01-preview' = {
-  name: 'dsrp-resources-redis-recipe'
+  name: 'dsrp-resources-redis-default-recipe'
   location: 'global'
   properties: {
     environment: env.id
     extensions: [
       {
         kind: 'kubernetesNamespace'
-        namespace: 'dsrp-resources-redis-recipe-app'
+        namespace: 'dsrp-resources-redis-default-recipe-app'
       }
     ]
   }
 }
 
 resource redis 'Applications.Datastores/redisCaches@2023-10-01-preview' = {
-  name: 'rds-recipe'
+  name: 'rds-default-recipe'
   location: 'global'
   properties: {
     environment: env.id
     application: app.id
-    recipe: {
-      name: 'rediscache'
-    }
   }
 }
