+++ test/functional-portable/datastoresrp/noncloud/resources/testdata/datastoresrp-resources-redis-recipe.bicep
@@ -1,28 +1,21 @@
 import radius as radius
 
-param scope string = resourceGroup().id
-
 param registry string
 
 param version string
 
 resource env 'Applications.Core/environments@2023-10-01-preview' = {
-  name: 'dsrp-resources-env-default-recipe-env'
+  name: 'dsrp-resources-env-recipe-env'
   location: 'global'
   properties: {
     compute: {
       kind: 'kubernetes'
       resourceId: 'self'
-      namespace: 'dsrp-resources-env-default-recipe-env'
-    }
-    providers: {
-      azure: {
-        scope: scope
-      }
+      namespace: 'dsrp-resources-env-recipe-env'
     }
     recipes: {
       'Applications.Datastores/redisCaches': {
-        default: {
+        rediscache: {
           templateKind: 'bicep'
           templatePath: '${registry}/test/testrecipes/test-bicep-recipes/redis-recipe-value-backed:${version}'
         }
@@ -32,24 +25,27 @@ resource env 'Applications.Core/environments@2023-10-01-preview' = {
 }
 
 resource app 'Applications.Core/applications@2023-10-01-preview' = {
-  name: 'dsrp-resources-redis-default-recipe'
+  name: 'dsrp-resources-redis-recipe'
   location: 'global'
   properties: {
     environment: env.id
     extensions: [
       {
         kind: 'kubernetesNamespace'
-        namespace: 'dsrp-resources-redis-default-recipe-app'
+        namespace: 'dsrp-resources-redis-recipe-app'
       }
     ]
   }
 }
 
 resource redis 'Applications.Datastores/redisCaches@2023-10-01-preview' = {
-  name: 'rds-default-recipe'
+  name: 'rds-recipe'
   location: 'global'
   properties: {
     environment: env.id
     application: app.id
+    recipe: {
+      name: 'rediscache'
+    }
   }
 }
